Refactor handlers
