+++ src/tui/handlers/split_config.rs
@@ -1,0 +1,6 @@
+use crossterm::event::KeyCode;
+use crate::tui::app::App;
+
+pub fn handle_split_config_input(_key: KeyCode, _app: &mut App) {
+    todo!("Implement split configuration input handling");
+}
